# app/src/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/authentication/sign-up": {
      "post": {
        "operationId": "AuthenticationController_singUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignUpDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Authentication"]
      }
    },
    "/api/authentication/sign-in": {
      "post": {
        "operationId": "AuthenticationController_singIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignInDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Authentication"]
      }
    }
  },
  "info": {
    "title": "Job Up - Swagger",
    "description": "Swagger for Job Up ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SignUpDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string", "minLength": 10 }
        },
        "required": ["email", "password"]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string", "minLength": 10 }
        },
        "required": ["email", "password"]
      }
    }
  }
}